pyproject.toml
[project]
name = "fastapi-celery-project"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "saitaro", email = "nort87@mail.ru" }
]
dependencies = [
    "fastapi==0.111.0",
    "uvicorn==0.25.0",
    "blue>=0.9.1",
    "celery==5.3.6",
    "redis>=5.0.8",
    "flower>=2.0.1",
    "sqlalchemy>=2.0.30",
    "alembic>=1.13.1",
    "psycopg2-binary==2.9.9",
    "watchfiles>=0.23.0",
    "jinja2>=3.1.4",
    "requests>=2.32.3",
    "celery-types>=0.22.0",
    "asgiref>=3.8.1",
    "asyncio-redis>=0.16.0",
    "broadcaster>=0.3.1",
    "websockets>=12.0",
    "python-socketio>=5.11.3",
]
readme = "README.md"
requires-python = ">= 3.11"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/fastapi_celery_project"]
